+++ sf_core/src/apis/database_driver_v1/statement.rs
@@ -584,6 +584,7 @@ impl DatabaseDriverV1 {
                     tls_config,
                     proxy_config,
                     put_fastfail,
+                    put_compress_level,
                     get_fastfail,
                 ) = {
                     let conn = conn.lock().await;
@@ -598,6 +599,7 @@ impl DatabaseDriverV1 {
                         put_fastfail_override
                             .or_else(|| conn.put_fastfail())
                             .unwrap_or(self.wrapper_presets.put_get_fastfail_default),
+                        conn.put_compress_level(self.wrapper_presets.put_compress_level_default),
                         get_fastfail_override
                             .or_else(|| conn.get_fastfail())
                             .unwrap_or(self.wrapper_presets.put_get_fastfail_default),
@@ -613,6 +615,7 @@ impl DatabaseDriverV1 {
                     use_s3_regional_url_session_param,
                     skip_upload_on_content_match,
                     put_fastfail,
+                    put_compress_level,
                     get_fastfail,
                     cwd,
                     unsafe_file_write,