+++ sf_core/tests/integration/http/s3_multipart.rs
@@ -180,6 +180,7 @@ async fn should_upload_and_download_via_s3_multipart_roundtrip() {
         legacy_odbc_compression_autodetect: false,
         skip_upload_on_content_match: false,
         multipart,
+        put_compress_level: 9,
     };
     let upload_result = upload_single_file(
         upload,